+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on Keep a Changelog, and this project follows semver for the
 
 ## [Unreleased]
 
+## [0.5.2] - 2026-05-14
+
 ### Added
 
 - Added `-g` as a shorthand for `--global` on `install`, `publish`, `agent use`, and `agent clear`.